NOT WORKING - pre context impl
This does work for small files or even a couple small files and a large
one (75GB) but eventually either a) the tick gets too large (8s) or
b) nothing seems to happen.

Suspect that the random hangs are happening when s3 gets sick of
returning RequestErrors and it just drops the connection without
response.

For this we should have a timeout to that each can be replayed and not
just stop the pipeline waiting for responses that may never come.